requirements.txt
tweepy==4.1.0
Pillow==8.3.1
requests==2.26.0
airtable-python-wrapper==0.15.0